% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{all_gen}
\alias{all_gen}
\title{Coverage and length results generation}
\usage{
all_gen(
  eval,
  cb.l,
  cb.u,
  h.t,
  h.c = h.t,
  reg.name = c("AK", "AK.bin"),
  reg.spec,
  M,
  cvar.spec = NULL,
  scale = NULL,
  x.spec
)
}
\arguments{
\item{eval}{a vector of evaluation points in the support of the regressor}

\item{cb.l}{a vector of lower confidence band values}

\item{cb.u}{a vector of upper confidence band values}

\item{h.t}{bandwidth used at each evaluation points for treated observations or for all observations
if there is no treated/control distinction.}

\item{h.c}{bandwidth used at each evaluation points for control observations;
it can be left unspecified if there is no treated/control distinction.}

\item{reg.name}{class of regression functions to be used; \code{"AK"} denotes regression functions
in Section 5 of Armstrong and Kolesár (2020); \code{"AK.bin"} denotes denotes the case where
\eqn{y = 1(f(x) + u > 0)}, using regression functions
in Section 5 of Armstrong and Kolesár (2020), where \eqn{u} is a normal random variable.}

\item{reg.spec}{a number denoting the type of regression function among the class of functions chosen
in \code{reg.name}.}

\item{M}{bound on the second derivative}

\item{cvar.spec}{specification for conditional variance function, one of \code{c("homo", "het.AK")}.}

\item{scale}{scale parameter, a positive constant multiplied to the noise.}

\item{x.spec}{distribution of the regressor, which determines the weights. Supports
the uniform distribution (\code{"unif"}) and the normalized beta(2, 2) distribution (\code{"beta"}),
both with the support of \eqn{[-1, 1]}.}
}
\value{
a vector of 1) coverage indicator, 2) weighted length, 3) supremum length,
4) a vector of confidence band lengths at each evaluation points and
5) a vector of bandwidths used at each evaluation points given a confidence band for treated and control observations,
6) a vector of evaluation points.
}
\description{
Generates 1) coverage indicator, 2) confidence band lengths at each evaluation points and 3) bandwidths used
at each evaluation points given a confidence band.
}
