% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChengChen.R
\name{CB_RBC}
\alias{CB_RBC}
\title{Confidence Band Using Bootstrap}
\usage{
CB_RBC(y, x, eval, B = 1000, level = 0.95, fixed = TRUE, bwselect = "mse-dpi")
}
\arguments{
\item{y}{dependent variable.}

\item{x}{independent variable.}

\item{eval}{vector of evaluation points.}

\item{B}{number of bootstrap simulations; default is \code{B = 1000}.}

\item{level}{confidence level used for confidence band; default is
\code{level = 0.95}.}

\item{fixed}{calculates the fixed length if \code{TRUE}; otherwise, CB is calculated
according to Remark 2 of Cheng and Chen (2019).}

\item{bwselect}{bandwidth selection method of \code{\link[nprobust]{lprobust}}; default is \code{"mse-dpi"}.}
}
\value{
a data frame containing index set and corresponding confidence band values
}
\description{
Genereates a confidence band using the procedure described in
Cheng and Chen (2019).
}
\details{
\code{B = 1000} is used in Cheng and Chen (2019).
}
\examples{
x <- stats::runif(500, min = -1, max = 1)
y <- x + rnorm(500, 0, 1/4)
eval <- seq(from = -1, to = 1, length.out = 20)
CB_RBC(y, x, eval, 2)
CB_RBC(y, x, eval, 2, fixed = FALSE)
}
\references{
{

\cite{Cheng,  G., and Y.-C. Chen. 2019.
"Nonparametric  inference  via  bootstrapping  the
debiased estimator."
Electronic Journal of Statistics, 13 (1): 2194–2256.}
}
}
