% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChengChen.R
\name{de_lpr}
\alias{de_lpr}
\title{Debiased local polynomial estimation.}
\usage{
de_lpr(X, Y, h, tau = 1, x_seq)
}
\arguments{
\item{X}{univariate X random variable.}

\item{Y}{univariate Y random variable.}

\item{h}{bandwidth for estimating the regression function.}

\item{tau}{ratio for bandwidths of estimating the derivative function over the bandwidth}

\item{x_seq}{x data point for evaluting the local polynomial estimator.}
}
\value{
a list with estimates with original local polynomial estimation (\code{lpr}),
and estimates with debiased local polynomial estimation (\code{lpr_de}).
}
\description{
A code provided by Gang Cheng at University of Washington (gangc at uw.edu)
}
\examples{
x <- stats::runif(500, min = -1, max = 1)
y <- x + rnorm(500, 0, 1/4)
eval <- seq(from = -0.9, to = 0.9, length.out = 20)
de_lpr(x, y, 0.1, x_seq = eval)
}
