% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{eps_gen}
\alias{eps_gen}
\title{Noise generation}
\usage{
eps_gen(n.sim, cvar.spec, dist = c("norm", "lnorm", "logis"), x, scale)
}
\arguments{
\item{n.sim}{number of simulation to use.}

\item{cvar.spec}{specification for conditional variance function, one of \code{c("homo", "het.AK")}.}

\item{dist}{distribution for noise function; choose among normal (\code{"norm"}),
log-normal distribution (\code{"lnorm"}) or logistic (\code{"logis"}).}

\item{x}{a vector of regressors}

\item{scale}{scale parameter, a positive constant multiplied to the noise.}
}
\value{
matrix of noise terms with \code{nrow = length(x)} and \code{ncol = n.sim}.
}
\description{
Noise generation
}
\examples{
x <- seq(-1, 1, length.out = 100)
eps_gen(150, "het.AK", "lnorm", x, 1/2)
}
