% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{f.BCR}
\alias{f.BCR}
\title{Berry, Caroll, and Ruppert (2002) Regression Function}
\usage{
f.BCR(x, M = 3 * pi/2)
}
\arguments{
\item{x}{a vector of x values.}

\item{M}{an additional parameter.}
}
\description{
Berry, Caroll, and Ruppert (2002) Regression Function
}
\examples{
x <- runif(100, -1, 1)
f.BCR(x)
}
\references{
{

\cite{S. M. Berry, R. J. Carroll, and D. Ruppert. 2002.
"Bayesian smoothing and regression splines for measurement error problems."
Journal of the American Statistical Association 97 (457): 160–169.}

}
}
