% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{f_AK}
\alias{f_AK}
\title{Regression functions in Armstrong and Kolesar (2020)}
\usage{
f_AK(spec, x, M)
}
\arguments{
\item{spec}{which specification to use; choose among {1,2,3}.}

\item{x}{a vector of x values.}

\item{M}{bound on the second derivative}
}
\description{
Functions \eqn{f_1(x)}, \eqn{f_2(x)}, and \eqn{f_3(x)} as defined in Section 5 of Armstrong and Kolesár (2020).
}
\examples{
x <- runif(100, -1, 1)
M <- 2
f_AK(1, x, M)
f_AK(2, x, M)
f_AK(3, x, M)
}
\references{
{

\cite{Armstrong, Timothy B., and Michal Kolesár. 2020.
"Simple and Honest Confidence Intervals in Nonparametric Regression."
Quantitative Economics 11 (1): 1–39.}

}
}
