% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{noncov_len}
\alias{noncov_len}
\title{Non-coverage "length" calculation}
\usage{
noncov_len(
  eval,
  cb.l,
  cb.u,
  reg.name = c("AK", "AK.bin"),
  reg.spec,
  M,
  cvar.spec = NULL,
  scale = NULL
)
}
\arguments{
\item{eval}{a vector of evaluation points in the support of the regressor}

\item{cb.l}{a vector of lower confidence band values}

\item{cb.u}{a vector of upper confidence band values}

\item{reg.name}{class of regression functions to be used; \code{"AK"} denotes regression functions
in Section 5 of Armstrong and Kolesár (2020); \code{"AK.bin"} denotes denotes the case where
\eqn{y = 1(f(x) + u > 0)}, using regression functions
in Section 5 of Armstrong and Kolesár (2020), where \eqn{u} is a normal random variable.}

\item{reg.spec}{a number denoting the type of regression function among the class of functions chosen
in \code{reg.name}.}

\item{M}{bound on the second derivative}

\item{cvar.spec}{specification for conditional variance function, one of \code{c("homo", "het.AK")}.}

\item{scale}{scale parameter, a positive constant multiplied to the noise.}
}
\value{
a vector of length \code{2 * len(eval)}, with a vector of lower non-coverage length
and a vector of upper non-coverage length
}
\description{
Non-coverage "length" calculation
}
