% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{obs_gen}
\alias{obs_gen}
\title{Observation generation}
\usage{
obs_gen(
  n,
  n.sim,
  reg.name = c("AK", "AK.bin"),
  reg.spec,
  M,
  cvar.spec,
  dist,
  scale,
  x.spec = c("unif", "beta"),
  x.l = -1,
  x.u = 1
)
}
\arguments{
\item{n}{number of observations}

\item{n.sim}{number of simulation to use.}

\item{reg.name}{class of regression functions to be used; \code{"AK"} denotes regression functions
in Section 5 of Armstrong and Kolesár (2020); \code{"AK.bin"} denotes denotes the case where
\eqn{y = 1(f(x) + u > 0)}, using regression functions
in Section 5 of Armstrong and Kolesár (2020), where \eqn{u} is a normal random variable.}

\item{reg.spec}{a number denoting the type of regression function among the class of functions chosen
in \code{reg.name}.}

\item{M}{bound on the second derivative}

\item{cvar.spec}{specification for conditional variance function, one of \code{c("homo", "het.AK")}.}

\item{dist}{distribution for noise function; either normal (\code{"norm"}) or log-normal distribution (\code{"lnorm"}).}

\item{scale}{scale parameter, a positive constant multiplied to the noise.}

\item{x.spec}{distribution of the regressor, either uniform distribution \code{"unif"} or beta distribution \code{"beta"}}

\item{x.l}{lower end of the support of the regressor.}

\item{x.u}{upper end of the support of the regressor.}
}
\value{
list of following components
\describe{
\item{x}{vector of regressors}
\item{y}{matrix of outcome variables with \code{nrow = n} and \code{ncol = n.sim}}
}
}
\description{
Observation generation
}
\examples{
obs_gen(100, 100, "AK.bin", 2, 1, "het.AK", "norm", 1/2, "beta")
}
