% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{s_fun}
\alias{s_fun}
\title{Function \eqn{s(x)} in Armstrong and Kolesar (2020)}
\usage{
s_fun(x)
}
\arguments{
\item{x}{a vector of x values}
}
\description{
Function \eqn{s(x)} as defined in Section 5 of Armstrong and Kolesár (2020).
}
\examples{
x <- runif(100, -1, 1)
s_fun(x)
}
\references{
{

\cite{Armstrong, Timothy B., and Michal Kolesár. 2020.
"Simple and Honest Confidence Intervals in Nonparametric Regression."
Quantitative Economics 11 (1): 1–39.}

}
}
