% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{weighted_len}
\alias{weighted_len}
\title{Weighted length of a confidence band}
\usage{
weighted_len(eval, cb.l, cb.u, x.spec = c("unif", "beta"))
}
\arguments{
\item{eval}{a vector of evaluation points in the support of the regressor}

\item{cb.l}{a vector of lower confidence band values}

\item{cb.u}{a vector of upper confidence band values}

\item{x.spec}{distribution of the regressor, which determines the weights. Supports
the uniform distribution (\code{"unif"}) and the normalized beta(2, 2) distribution (\code{"beta"}),
both with the support of \eqn{[-1, 1]}.}
}
\value{
a scalar value of weighted length.
}
\description{
Computes weighted length of a confidence band, where the weights are given
by the design of the regressor.
}
