% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{y_gen}
\alias{y_gen}
\title{Outcome generation}
\usage{
y_gen(
  n.sim,
  reg.name = c("AK", "AK.bin"),
  reg.spec,
  M,
  x,
  cvar.spec,
  dist,
  scale
)
}
\arguments{
\item{n.sim}{number of simulation to use.}

\item{reg.name}{class of regression functions to be used; \code{"AK"} denotes regression functions
in Section 5 of Armstrong and Kolesár (2020); \code{"AK.bin"} denotes denotes the case where
\eqn{y = 1(f(x) + u > 0)}, using regression functions
in Section 5 of Armstrong and Kolesár (2020), where \eqn{u} is a normal random variable.}

\item{reg.spec}{a number denoting the type of regression function among the class of functions chosen
in \code{reg.name}.}

\item{M}{bound on the second derivative}

\item{x}{a vector of regressors}

\item{cvar.spec}{specification for conditional variance function, one of \code{c("homo", "het.AK")}.}

\item{dist}{distribution for noise function; either normal (\code{"norm"}) or log-normal distribution (\code{"lnorm"}).}

\item{scale}{scale parameter, a positive constant multiplied to the noise.}
}
\value{
matrix of outcome variables with \code{nrow = length(x)} and \code{ncol = n.sim}.
}
\description{
Outcome generation
}
\examples{
x <- seq(-1, 1, length.out = 50)
y_gen(70, "AK", 3, 1, x, "het.AK", "lnorm", 1/2)
y_gen(70, "AK.bin", 2, 1, x, "het.AK", "norm", 1/2)
}
